+++ src/components/bookDetails/__tests__/FulfillmentCard.test.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { render, waitForElementToBeRemoved, waitFor } from "test-utils";
-import { mergeBook } from "test-utils/fixtures";
+import { mergeBook, mockSetBook } from "test-utils/fixtures";
 import FulfillmentCard from "../FulfillmentCard";
 import userEvent from "@testing-library/user-event";
 import _download from "downloadjs";
@@ -15,11 +15,19 @@ import {
 import { ProblemDocument } from "types/opds1";
 import fetchMock from "jest-fetch-mock";
 import { mockPush } from "test-utils/mockNextRouter";
+import * as fetch from "dataflow/opds1/fetch";
+import { ServerError } from "errors";
 import * as env from "utils/env";
 
 jest.mock("downloadjs");
 window.open = jest.fn();
 
+jest.mock("dataflow/opds1/fetch");
+
+(fetch as any).fetchBook = jest.fn();
+const mockFetchBook = fetch.fetchBook as jest.MockedFunction<
+  typeof fetch.fetchBook
+>;
 /**
  * Borrowable
  * OnHold
@@ -46,6 +54,12 @@ describe("BorrowableBook", () => {
   });
 
   test("borrow button fetches book and displays borrow errors", async () => {
+    const problem: ProblemDocument = {
+      detail: "Can't do that",
+      title: "Nope",
+      status: 418
+    };
+    mockFetchBook.mockRejectedValue(new ServerError("/borrow", 418, problem));
     const utils = render(<FulfillmentCard book={borrowableBook} />, {
       user: { isAuthenticated: true }
     });
@@ -57,8 +71,7 @@ describe("BorrowableBook", () => {
 
     // the borrow button should be gone now
     await waitForElementToBeRemoved(() => utils.getByText("Borrowing..."));
-    // there is an error because we didn't mock fetch to return something
-    expect(utils.getByText("Error: An unknown error occurred."));
+    expect(utils.getByText("Error: Can't do that"));
   });
 });
 
@@ -79,6 +92,12 @@ describe("OnHoldBook", () => {
   });
 
   test("borrow button fetches url and shows error", async () => {
+    const problem: ProblemDocument = {
+      detail: "Can't do that",
+      title: "Nope",
+      status: 418
+    };
+    mockFetchBook.mockRejectedValue(new ServerError("/borrow", 418, problem));
     const utils = render(<FulfillmentCard book={onHoldBook} />, {
       user: { isAuthenticated: true }
     });
@@ -90,8 +109,7 @@ describe("OnHoldBook", () => {
 
     // the borrow button should be gone now
     await waitForElementToBeRemoved(() => utils.getByText("Borrowing..."));
-    // there is an error because we didn't mock fetch to return something
-    expect(utils.getByText("Error: An unknown error occurred."));
+    expect(utils.getByText("Error: Can't do that"));
   });
 
   test("handles lack of availability.until info", () => {
@@ -169,6 +187,12 @@ describe("ReservableBook", () => {
   });
 
   test("shows reserve button which fetches book", async () => {
+    const problem: ProblemDocument = {
+      detail: "Can't do that",
+      title: "Nope",
+      status: 418
+    };
+    mockFetchBook.mockRejectedValue(new ServerError("/borrow", 418, problem));
     const utils = render(<FulfillmentCard book={reservableBook} />, {
       user: { isAuthenticated: true }
     });
@@ -180,8 +204,7 @@ describe("ReservableBook", () => {
 
     // the borrow button should be gone now
     await waitForElementToBeRemoved(() => utils.getByText("Reserving..."));
-    // there is an error because we didn't mock fetch to return something
-    expect(utils.getByText("Error: An unknown error occurred."));
+    expect(utils.getByText("Error: Can't do that"));
   });
 });
 
@@ -198,11 +221,62 @@ describe("reserved", () => {
     }
   });
 
-  test("displays disabled reserve button", () => {
+  test("displays cancel reservation button that calls apropriate url", async () => {
+    const unreservedBook = mergeBook<BorrowableBook>({
+      status: "borrowable",
+      borrowUrl: "/borrow"
+    });
+    mockFetchBook.mockResolvedValue(unreservedBook);
     const utils = render(<FulfillmentCard book={reservedBook} />);
-    const reserveButton = utils.getByRole("button", { name: "Reserved" });
-    expect(reserveButton).toBeInTheDocument();
-    expect(reserveButton).toBeDisabled();
+    const revokeButton = utils.getByRole("button", {
+      name: "Cancel Reservation"
+    });
+    expect(revokeButton).toBeInTheDocument();
+
+    userEvent.click(revokeButton);
+
+    expect(
+      await utils.findByRole("button", { name: "Cancelling..." })
+    ).toBeInTheDocument();
+
+    expect(mockFetchBook).toHaveBeenCalledWith(
+      "/revoke",
+      "http://test-cm.com/catalogUrl",
+      "user-token"
+    );
+
+    expect(mockSetBook).toHaveBeenCalledWith(unreservedBook, reservedBook.id);
+  });
+
+  test("handles cancel reservation errors", async () => {
+    const problem: ProblemDocument = {
+      detail: "Can't do that",
+      title: "Nope",
+      status: 418
+    };
+    mockFetchBook.mockRejectedValue(new ServerError("/revoke", 418, problem));
+    const utils = render(<FulfillmentCard book={reservedBook} />);
+    const revokeButton = utils.getByRole("button", {
+      name: "Cancel Reservation"
+    });
+    expect(revokeButton).toBeInTheDocument();
+
+    userEvent.click(revokeButton);
+
+    expect(
+      await utils.findByRole("button", { name: "Cancelling..." })
+    ).toBeInTheDocument();
+
+    expect(mockFetchBook).toHaveBeenCalledWith(
+      "/revoke",
+      "http://test-cm.com/catalogUrl",
+      "user-token"
+    );
+
+    expect(await utils.findByText("Error: Can't do that")).toBeInTheDocument();
+    expect(
+      await utils.findByRole("button", { name: "Cancel Reservation" })
+    ).toBeInTheDocument();
   });
 
   test("displays number of patrons in queue and your position", () => {
@@ -264,6 +338,33 @@ describe("FulfillableBook", () => {
       status: "available",
       until: "2020-06-18"
     }
+  });
+
+  test("displays return button that calls apropriate url", async () => {
+    const unborrowed = mergeBook<BorrowableBook>({
+      status: "borrowable",
+      borrowUrl: "/borrow"
+    });
+    mockFetchBook.mockResolvedValue(unborrowed);
+    const utils = render(<FulfillmentCard book={downloadableBook} />);
+    const revokeButton = utils.getByRole("button", {
+      name: "Return"
+    });
+    expect(revokeButton).toBeInTheDocument();
+
+    userEvent.click(revokeButton);
+
+    expect(
+      await utils.findByRole("button", { name: "Returning..." })
+    ).toBeInTheDocument();
+
+    expect(mockFetchBook).toHaveBeenCalledWith(
+      "/revoke",
+      "http://test-cm.com/catalogUrl",
+      "user-token"
+    );
+
+    expect(mockSetBook).toHaveBeenCalledWith(unborrowed, downloadableBook.id);
   });
 
   test("constructs link to viewer for OpenAxis Books", () => {
@@ -377,9 +478,7 @@ describe("FulfillableBook", () => {
     const utils = render(<FulfillmentCard book={bookWithRedirect} />);
 
     expect(utils.queryByText("Ready to Read!")).not.toBeInTheDocument();
-    expect(
-      utils.getByText("You're ready to read this book in SimplyE!")
-    ).toBeInTheDocument();
+    expect(utils.getByText("Ready to read in SimplyE!")).toBeInTheDocument();
     expect(utils.getByText("You have this book on loan.")).toBeInTheDocument();
     expect(
       utils.getByText("If you would rather read on your computer, you can:")
@@ -391,7 +490,7 @@ describe("FulfillableBook", () => {
     mockConfig({ companionApp: "openebooks" });
     const utils = render(<FulfillmentCard book={bookWithRedirect} />);
     expect(
-      utils.getByText("You're ready to read this book in Open eBooks!")
+      utils.getByText("Ready to read in Open eBooks!")
     ).toBeInTheDocument();
     expect(utils.getByText("You have this book on loan.")).toBeInTheDocument();
   });