+++ src/dataflow/opds1/parse.ts
@@ -25,7 +25,7 @@ import {
   FulfillmentLink,
   AnyBook
 } from "interfaces";
-import { IncorrectAdeptMediaType } from "types/opds1";
+import { IncorrectAdobeDrmMediaType } from "types/opds1";
 import { getAppSupportLevel } from "utils/fulfill";
 import { TrackOpenBookRel } from "types/opds1";
 
@@ -76,9 +76,13 @@ const resolve = (base: string, relative: string) =>
  * Adobe media types.
  */
 export function fixMimeType(
-  mimeType: OPDS1.IndirectAcquisitionType | typeof OPDS1.IncorrectAdeptMediaType
+  mimeType:
+    | OPDS1.IndirectAcquisitionType
+    | typeof OPDS1.IncorrectAdobeDrmMediaType
 ): OPDS1.IndirectAcquisitionType {
-  return mimeType === IncorrectAdeptMediaType ? OPDS1.AdeptMediaType : mimeType;
+  return mimeType === IncorrectAdobeDrmMediaType
+    ? OPDS1.AdobeDrmMediaType
+    : mimeType;
 }
 
 function parseFormat(format: OPDSAcquisitionLink | OPDSIndirectAcquisition) {
@@ -113,12 +117,37 @@ function buildFulfillmentLink(feedUrl: string) {
       indirectionType: fixMimeType(
         indirectionType as
           | OPDS1.IndirectAcquisitionType
-          | typeof OPDS1.IncorrectAdeptMediaType
+          | typeof OPDS1.IncorrectAdobeDrmMediaType
       )
     };
   };
 }
 
+/**
+ * A book cannot be returned (even with a revoke url) if:
+ *  - It is locked in to Adobe ACS
+ *  - It is an axisnow book
+ *  - It comes from certain vendors
+ *
+ * We use a heuristic to determine when books can be returned. This may
+ * need to be updated in the future.
+ */
+function canReturnFulfillableBook(links: OPDSAcquisitionLink[]): boolean {
+  return !!links.map(parseFormat).find(link => {
+    // no AxisNow
+    if (link.contentType === OPDS1.AxisNowWebpubMediaType) return false;
+    // match if there is otherwise a direct link. These won't be present
+    // if the book has been locked in to Adobe ACS
+    if (!link.indirectionType) return true;
+    // match if it has a link to read online externally.
+    if (link.contentType === OPDS1.ExternalReaderMediaType) return true;
+    return false;
+  });
+}
+
+function findRevokeUrl(links: OPDSLink[]) {
+  return links.find(link => link.rel === OPDS1.RevokeLinkRel)?.href ?? null;
+}
 /**
  * HTML Sanitizer
  */
@@ -211,8 +240,7 @@ export function entryToBook(entry: OPDSEntry, feedUrl: string): AnyBook {
     link => link.supportLevel !== "unsupported"
   );
 
-  const revokeUrl =
-    entry.links.find(link => link.rel === OPDS1.RevokeLinkRel)?.href ?? null;
+  const revokeUrl = findRevokeUrl(entry.links);
 
   const trackOpenBookLink = entry.links.find(isTrackOpenBookLink);
 
@@ -253,13 +281,17 @@ export function entryToBook(entry: OPDSEntry, feedUrl: string): AnyBook {
       ...supportedFulfillmentLinks,
       ...openAccessLinks
     ];
+    if (book.title === "Apollo and the Battle of the Birds") {
+      console.log(acquisitionLinks);
+    }
     return {
       ...book,
       status: "fulfillable",
       fulfillmentLinks: allFulfillmentLinks,
-      revokeUrl
+      revokeUrl: canReturnFulfillableBook(acquisitionLinks) ? revokeUrl : null
     };
   }
+
   // it's a reserved book
   if (availability?.status === "reserved") {
     return {
@@ -268,6 +300,7 @@ export function entryToBook(entry: OPDSEntry, feedUrl: string): AnyBook {
       revokeUrl
     };
   }
+
   // it's a reservable book
   if (borrowLink && borrowLink.availability.status === "unavailable") {
     return {