+++ src/types/opds1.ts
@@ -56,9 +56,9 @@ export const OPDSEntryMediaType =
  * decrypted by us, so just download them and let the user open it in an app
  * that can deal with it.
  */
-export const AdeptMediaType = "application/vnd.adobe.adept+xml";
+export const AdobeDrmMediaType = "application/vnd.adobe.adept+xml";
 // there was an issue with incorrect Adept media types being sent.
-export const IncorrectAdeptMediaType = "vnd.adobe/adept+xml";
+export const IncorrectAdobeDrmMediaType = "vnd.adobe/adept+xml";
 // this is not yet supported, but the description is here:
 // https://github.com/NYPL-Simplified/Simplified/wiki/OPDSForDistributors#bearer-token-propagation
 export const BearerTokenMediaType =
@@ -66,7 +66,7 @@ export const BearerTokenMediaType =
 
 export type IndirectAcquisitionType =
   | typeof OPDSEntryMediaType
-  | typeof AdeptMediaType
+  | typeof AdobeDrmMediaType
   | typeof BearerTokenMediaType;
 
 /**